+++ build.sh
@@ -1538,7 +1538,7 @@ x.y.z Release notes (yyyy-MM-dd)
 * Realm Object Server: 3.21.0 or later.
 * Realm Studio: 3.12 or later.
 * APIs are backwards compatible with all previous releases in the 5.x.y series.
-* Carthage release for Swift is built with Xcode 11.7.
+* Carthage release for Swift is built with Xcode 12.
 
 ### Internal
 * Upgraded realm-core from ? to ?