+++ Package.swift
@@ -4,7 +4,7 @@ import PackageDescription
 import Foundation
 
 let coreVersionStr = "6.1.1"
-let cocoaVersionStr = "5.4.3"
+let cocoaVersionStr = "5.4.4"
 
 let coreVersionPieces = coreVersionStr.split(separator: ".")
 let coreVersionExtra = coreVersionPieces[2].split(separator: "-")