+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-x.y.z Release notes (yyyy-MM-dd)
+5.4.4 Release notes (2020-09-25)
 =============================================================
+
 ### Enhancements
 * Improve the asymtotic performance of NOT IN queries on indexed properties. It
   is now O(Number of Rows) rather than O(Number of Rows \* Number of values in IN clause.)
 
 ### Fixed
+
 * Fix a crash inside `realm::Array(Type)::init_from_mem()` which would
   sometimes occur when running a query over links immediately after creating
   objects of the queried type.
@@ -13,9 +15,10 @@ x.y.z Release notes (yyyy-MM-dd)
   from optional to non-optional.
 
 ### Compatibility
+
 * File format: Generates Realms with format v11 (Reads and upgrades all previous formats)
 * Realm Object Server: 3.21.0 or later.
-* Realm Studio: 3.12 or later.
+* Realm Studio: 5.0.0 or later.
 * APIs are backwards compatible with all previous releases in the 5.x.y series.
 * Carthage release for Swift is built with Xcode 12.